Implement voting-scheme for device challenges
Voting scheme based on PLCR-Voting for device challenging. Allows users
to challenge devices, vote on those challenges (with WEEV as a stake)
and resolve challenges of finished votings. Only registries that are
spawned through a valid factory are able to initiate votings. Details
on the voting process and its token mechanisms are included in the
token paper.