Implement voting for registries and marketplaces, renamed weeveFactory
Implemented the voting/challenging mechanism into the weeveNetwork
contract, allowing users to challenge registries and marketplaces,
vote on ongoing challenges and resolve them in the end.
Renamed the weeveFactory contract to weeveNetwork to reflect the current
functionalities.
Added a deactivate-function to the registries and marketplaces, which is
called after a successful challenge was resolved to deactivate the
corresponding registry or marketplace.
Updated the README according to the latest changes.